requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
dj-database-url==1.2.0
Django==4.1.7
django-rest-multiple-models==2.1.3
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-jsonapi==6.0.0
django-extensions==3.2.1
django-cors-headers==3.14.0
inflection==0.5.1
gunicorn==20.1.0
httplib2==0.14.0
incremental==16.10.1
netifaces==0.10.4
psycopg2-binary==2.9.5
pyasn1==0.4.2
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.2.0
Brotli==1.0.9
